chore: Moved analysis readers
