+++ ra2ce/analyses/configuration/readers/analysis_config_reader_base.py
@@ -22,17 +22,20 @@
 
 from pathlib import Path
 
-from ra2ce.configuration.readers.ini_config_reader_base import (
-    IniConfigurationReaderBase,
+from ra2ce.common.configuration.ini_configuration_reader_protocol import (
+    IniConfigurationReaderProtocol,
 )
 from ra2ce.configuration.validators.ini_config_validator_base import (
     DirectAnalysisNameList,
     IndirectAnalysisNameList,
 )
 from ra2ce.common.io.readers.ini_file_reader import IniFileReader
+import logging
+from pathlib import Path
+from shutil import copyfile
 
 
-class AnalysisConfigReaderBase(IniConfigurationReaderBase):
+class AnalysisConfigReaderBase(IniConfigurationReaderProtocol):
     def _convert_analysis_types(self, config: dict) -> dict:
         def set_analysis_values(config_type: str):
             if config_type in config:
@@ -63,3 +66,62 @@ class AnalysisConfigReaderBase(IniConfigurationReaderBase):
         _config["input"] = config_path.parent / "input"
         _config["static"] = config_path.parent / "static"
         return _config
+
+    def _copy_output_files(self, from_path: Path, config_data: dict) -> None:
+        self._create_config_dir("output", config_data)
+        try:
+            copyfile(from_path, config_data["output"] / "{}.ini".format(from_path.stem))
+        except FileNotFoundError as e:
+            logging.warning(e)
+
+    def _create_config_dir(self, dir_name: str, config_data: dict):
+        _dir = config_data["root_path"] / config_data["project"]["name"] / dir_name
+        if not _dir.exists():
+            _dir.mkdir(parents=True)
+        config_data[dir_name] = _dir
+
+    def _parse_path_list(
+        self, property_name: str, path_list: str, config_data: dict
+    ) -> list[Path]:
+        _list_paths = []
+        for path_value in path_list.split(","):
+            path_value = Path(path_value)
+            if path_value.is_file():
+                _list_paths.append(path_value)
+                continue
+
+            _project_name_dir = (
+                config_data["root_path"] / config_data["project"]["name"]
+            )
+            abs_path = _project_name_dir / "static" / property_name / path_value
+            try:
+                assert abs_path.is_file()
+            except AssertionError:
+                abs_path = _project_name_dir / "input" / property_name / path_value
+
+            _list_paths.append(abs_path)
+        return _list_paths
+
+    def _update_path_values(self, config_data: dict) -> None:
+        """
+        TODO: Work in progress, for now it's happening during validation, which should not be the case.
+
+        Args:
+            config_data (dict): _description_
+        """
+        _file_types = {
+            "polygon": "network",
+            "hazard_map": "hazard",
+            "origins": "network",
+            "destinations": "network",
+            "locations": "network",
+        }
+        for config_header, value_dict in config_data.items():
+            if not isinstance(value_dict, dict):
+                continue
+            for header_prop, prop_value in value_dict.items():
+                _prop_name = _file_types.get(header_prop, None)
+                if _prop_name and prop_value:
+                    config_data[config_header][header_prop] = self._parse_path_list(
+                        _prop_name, prop_value, config_data
+                    )
